requirements.txt
# Web framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6

# Twilio
twilio==8.10.3

# OpenAI
openai==1.3.7

# WebSocket a asynchronní zpracování
websockets==12.0
uvicorn==0.27.1
websocket-client==1.7.0
gevent==23.9.1

# Databáze
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Utility
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
gunicorn==21.2.0
APScheduler==3.10.4

# Logování
structlog==24.1.0

# Cloud SQL
cloud-sql-python-connector==1.7.0
pg8000==1.30.5

# Google Cloud
google-cloud-logging==3.9.0
google-cloud-storage==2.14.0

# Testování
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# Vývojové nástroje
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Additional dependencies
WTForms==3.1.1
python-socketio==5.10.0
python-engineio==4.7.1 